pyproject.toml
# =============================================================================
# pyproject.toml
# =============================================================================
# This is like package.json but for Python.
# It tells Python WHAT packages your project needs.
#
# When someone does "pip install -e ." it reads this file
# and installs everything listed under [dependencies].
# =============================================================================

[project]
name = "resume-screening-env"
version = "0.1.0"
description = "OpenEnv Resume Screening Environment — AI agent training for HR tasks"
readme = "README.md"
requires-python = ">=3.10"             # Need Python 3.10 or newer

dependencies = [
    # OpenEnv core — gives us Environment, Action, Observation base classes
    "openenv-core>=0.2.1",

    # FastAPI — the web framework that creates our HTTP server
    "fastapi>=0.104.0",

    # Uvicorn — the server that RUNS FastAPI (like Node.js for Python)
    "uvicorn>=0.24.0",

    # Pydantic — for defining typed data models (Action, Observation classes)
    "pydantic>=2.0.0",

    # For the baseline agent script that calls OpenAI
    "openai>=1.0.0",

    # For computing rank correlation in Task 2 grader
    "scipy>=1.11.0",

    # Websockets — OpenEnv uses WebSocket connections
    "websockets>=12.0",
]

[project.optional-dependencies]
# Extra packages only needed during development/testing
dev = [
    "pytest>=7.0.0",
    "httpx>=0.24.0",         # For testing FastAPI endpoints
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends.legacy:build"

[tool.setuptools.packages.find]
# Tells setuptools where to find Python packages
where = ["."]
include = ["resume_screening_env*"]
